pyproject.toml
[tool.isort]
profile = "black"

[tool.pylint.messages_control]
disable = ["import-error"]

[tool.poetry]
name = "caspian"
version = "0.1.0"
description = "Take control of your music. A fully-featured, self-hosted, open-source music service."
authors = ["Nate <minecraftcrusher100@gmail.com>"]
license = "LGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
spotdl = "^3.9.6"
youtube_dl = "^2021.12.17"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-multipart = "^0.0.5"
storage3 = "^0.3.4"
postgrest-py = "^0.10.2"
audio-metadata = "^0.11.1"
supabase = "^0.5.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
